multi_tool/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
dnspython>=2.2.0
python-whois>=0.8.0
tabulate>=0.9.0
pyOpenSSL>=23.0.0
cryptography>=38.0.0
validators>=0.20.0
tldextract>=3.4.0
websocket-client>=1.5.0
selenium>=4.9.0
pymisp>=2.4.170
PyJWT>=2.6.0
python-dateutil>=2.8.2
email-validator>=2.0.0
zxcvbn>=4.4.28
trafilatura>=1.5.0